+++ lib/ui/profile/ProfileInfoPreviewPage.dart
@@ -28,11 +28,11 @@ class ProfileInfoPreviewPage extends StatefulWidget {
 
 class ProfileInfoPreviewPageState extends State<ProfileInfoPreviewPage> {
 
-  Auth2UserProfile? _profile;
+  Auth2UserProfile? _publicProfile;
 
   @override
   void initState() {
-    _profile = widget.profile?.buildPublic(widget.privacy, permitted: { Auth2FieldVisibility.public });
+    _publicProfile = widget.profile?.buildPublic(widget.privacy, permitted: { Auth2FieldVisibility.public });
     super.initState();
   }
 
@@ -61,7 +61,7 @@ class ProfileInfoPreviewPageState extends State<ProfileInfoPreviewPage> {
       )
     ]);
 
-  String? get _photoImageUrl => StringUtils.isNotEmpty(_profile?.photoUrl) ?
+  String? get _photoImageUrl => StringUtils.isNotEmpty(_publicProfile?.photoUrl) ?
     Content().getUserPhotoUrl(type: UserProfileImageType.medium, params: DirectoryProfilePhotoUtils.tokenUrlParam(widget.photoImageToken)) : null;
 
   double get _photoImageSize => MediaQuery.of(context).size.width / 3;
@@ -77,7 +77,7 @@ class ProfileInfoPreviewPageState extends State<ProfileInfoPreviewPage> {
           ),
         ],),
         Padding(padding: EdgeInsets.only(top: 12, bottom: 12), child:
-          DirectoryProfileDetails(_profile)
+          DirectoryProfileDetails(_publicProfile)
         ),
         //if (_canShare)
         //  _shareButton,
@@ -85,58 +85,22 @@ class ProfileInfoPreviewPageState extends State<ProfileInfoPreviewPage> {
   );
 
   Widget get _cardContentHeading => Column(children: [
-    Padding(padding: EdgeInsets.only(top: (_profile?.pronunciationUrl?.isNotEmpty == true) ? 0 : 12), child:
+    Padding(padding: EdgeInsets.only(top: (_publicProfile?.pronunciationUrl?.isNotEmpty == true) ? 0 : 12), child:
       RichText(textAlign: TextAlign.center, text: TextSpan(style: nameTextStyle, children: [
-        TextSpan(text: _profile?.fullName ?? ''),
-        if (_profile?.pronunciationUrl?.isNotEmpty == true)
+        TextSpan(text: _publicProfile?.fullName ?? ''),
+        if (_publicProfile?.pronunciationUrl?.isNotEmpty == true)
           WidgetSpan(alignment: PlaceholderAlignment.middle, child:
             DirectoryPronunciationButton(
-              url: _profile?.pronunciationUrl,
+              url: _publicProfile?.pronunciationUrl,
               data: widget.pronunciationAudioData,
               padding: EdgeInsets.symmetric(horizontal: 13, vertical: 12),
             ),
           ),
       ])),
     ),
-  if (_profile?.pronouns?.isNotEmpty == true)
-    Text(_profile?.pronouns ?? '', style: Styles().textStyles.getTextStyle('widget.detail.small'), textAlign: TextAlign.center,),
+  if (_publicProfile?.pronouns?.isNotEmpty == true)
+    Text(_publicProfile?.pronouns ?? '', style: Styles().textStyles.getTextStyle('widget.detail.small'), textAlign: TextAlign.center,),
   ],);
-
-  // bool get _canShare => (widget.onboarding == false) && Auth2().isOidcLoggedIn && (_profile?.isNotEmpty == true);
-
-  /* Widget get _shareButton => Row(children: [
-    Padding(padding: EdgeInsets.only(right: 4), child:
-      Styles().images.getImage('share', size: 14) ?? Container()
-    ),
-    Expanded(child:
-      LinkButton(
-        title: Localization().getStringEx('panel.profile.info.command.link.share.text', 'Share My Info'),
-        textStyle: Styles().textStyles.getTextStyle('widget.button.title.small.underline'),
-        textAlign: TextAlign.left,
-        padding: EdgeInsets.symmetric(vertical: 16),
-        onTap: _onShare,
-      ),
-    ),
-  ],); */
-
-  /* void _onShare() {
-    Analytics().logSelect(target: 'Share');
-    /*ProfileInfoShareSheet.present(context,
-      profile: _profile,
-      photoImageData: widget.photoImageData,
-      pronunciationAudioData: widget.pronunciationAudioData,
-    );*/
-    NotificationService().notify(ProfileHomePanel.notifySelectContent, [
-      ProfileContentType.share,
-      <String, dynamic>{
-        ProfileInfoSharePage.profileResultKey : ProfileInfoLoadResult(
-          profile: _profile,
-          photoImageData: widget.photoImageData,
-          pronunciationAudioData: widget.pronunciationAudioData,
-        )
-      }
-    ]);
-  } */
 
   TextStyle? get nameTextStyle =>
     Styles().textStyles.getTextStyleEx('widget.title.medium_large.fat', fontHeight: 0.85, textOverflow: TextOverflow.ellipsis);